+++ src/main/java/org/opensearch/security/ssl/DefaultSecurityKeyStore.java
@@ -870,24 +870,15 @@ public class DefaultSecurityKeyStore implements SecurityKeyStore {
         final SslProvider sslProvider,
         final ClientAuth authMode
     ) throws SSLException {
+        final SslContextBuilder _sslContextBuilder = AccessController.doPrivilegedChecked(
+            () -> configureSSLServerContextBuilder(SslContextBuilder.forServer(_key, _cert), sslProvider, ciphers, authMode)
+        );
 
-        try {
-            final SslContextBuilder _sslContextBuilder = AccessController.doPrivilegedChecked(
-                () -> configureSSLServerContextBuilder(SslContextBuilder.forServer(_key, _cert), sslProvider, ciphers, authMode)
-            );
-
-            if (_trustedCerts != null && _trustedCerts.length > 0) {
-                _sslContextBuilder.trustManager(_trustedCerts);
-            }
-
-            return buildSSLContext0(_sslContextBuilder);
-        } catch (final Exception e) {
-            if (e.getCause() instanceof SSLException) {
-                throw (SSLException) e.getCause();
-            } else {
-                throw new RuntimeException(e);
-            }
+        if (_trustedCerts != null && _trustedCerts.length > 0) {
+            _sslContextBuilder.trustManager(_trustedCerts);
         }
+
+        return buildSSLContext0(_sslContextBuilder);
     }
 
     private SslContext buildSSLServerContext(
@@ -899,23 +890,15 @@ public class DefaultSecurityKeyStore implements SecurityKeyStore {
         final SslProvider sslProvider,
         final ClientAuth authMode
     ) throws SSLException {
-        try {
-            final SslContextBuilder _sslContextBuilder = AccessController.doPrivilegedChecked(
-                () -> configureSSLServerContextBuilder(SslContextBuilder.forServer(_cert, _key, pwd), sslProvider, ciphers, authMode)
-            );
+        final SslContextBuilder _sslContextBuilder = AccessController.doPrivilegedChecked(
+            () -> configureSSLServerContextBuilder(SslContextBuilder.forServer(_cert, _key, pwd), sslProvider, ciphers, authMode)
+        );
 
-            if (_trustedCerts != null) {
-                _sslContextBuilder.trustManager(_trustedCerts);
-            }
-
-            return buildSSLContext0(_sslContextBuilder);
-        } catch (final Exception e) {
-            if (e.getCause() instanceof SSLException) {
-                throw (SSLException) e.getCause();
-            } else {
-                throw new RuntimeException(e);
-            }
+        if (_trustedCerts != null) {
+            _sslContextBuilder.trustManager(_trustedCerts);
         }
+
+        return buildSSLContext0(_sslContextBuilder);
     }
 
     private SslContextBuilder configureSSLServerContextBuilder(